Some initial work on Push and PushRelativeValue.cs
PushRelativeValue not working at present: symbol definitions are not accurate.

Added properties so that it's possible to enquire the locations from a RelativeBinding.

Adjusted the descriptions in Operation to be a bit clearer for PushRelativeValue and PushContext.

+++ trunk/TameScheme/Scheme/Compiler/BOp/Push.cs
@@ -28,7 +28,7 @@ namespace Tame.Scheme.Compiler.BOp
             }
             else if (op.a is long)
             {
-                il.Emit(OpCodes.Ldc_I8, (int)op.a);
+                il.Emit(OpCodes.Ldc_I8, (long)op.a);
             }
             else if (op.a is string)
             {
